requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data sources
yfinance>=0.2.40
fredapi>=0.5.1
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data validation
pandera>=0.17.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Progress bars
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Optional: for future ML components
# scikit-learn>=1.3.0
# statsmodels>=0.14.0
